d3d12 drawing vertex buffer correctly

+++ Platforms/Shared/Orbital.Video.D3D12.Native/VertexBuffer.cpp
@@ -9,7 +9,7 @@ extern "C"
 		return handle;
 	}
 
-	ORBITAL_EXPORT int Orbital_Video_D3D12_VertexBuffer_Init(VertexBuffer* handle, void* vertices, uint64_t vertexCount, uint32_t vertexSize)
+	ORBITAL_EXPORT int Orbital_Video_D3D12_VertexBuffer_Init(VertexBuffer* handle, void* vertices, uint32_t vertexCount, uint32_t vertexSize)
 	{
 		uint64_t bufferSize = vertexSize * vertexCount;
